Update Frameworks 2 Final Project.Rmd
This is the code for my final project in my Applied Analytics Frameworks and Methods II course. I worked with two partners to read a government based dataset on grant projects where the target variable was whether the grant was approved for funding or not. The project goal being the ability to determine whether we could predict approval/rejection of a grant based on a combination of structured and unstructured data points, with greater focused applied to the text based features.